requirement.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0  
colorama==0.4.6
decorator==4.4.2
faiss-cpu==1.7.4
filelock==3.16.1
fsspec==2024.10.0
google==3.0.0
google-api-core==2.23.0    
google-auth==2.36.0        
google-cloud-speech==2.28.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
huggingface-hub==0.26.3
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pillow==11.0.0
pocketsphinx==5.0.3
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
six==1.17.0
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.11.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.0.1
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.2.3
whisper==1.1.10